Backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
numpy==1.24.3
pillow==10.1.0
python-dotenv==1.0.0
werkzeug==3.0.1
